+++ src/main/java/org/bytesync/hotelmanagement/repository/GuestRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.jpa.repository.Query;
 import java.util.List;
 import java.util.Optional;
 
-public interface GuestRepository extends JpaRepository<Guest, Integer>, JpaSpecificationExecutor<Guest> {
+public interface GuestRepository extends JpaRepository<Guest, Long>, JpaSpecificationExecutor<Guest> {
     boolean existsByEmail(String email);
 
     boolean existsByPassport(String passport);
